requirements.txt
aiogram==2.25.1
asyncpg==0.27.0
authentication==1.1.0
axios==0.1.0
brotli==1.0.9
channels==4.0.0
clickuz==0.3
crispy-bootstrap5==0.7
data==0.4
dj-rest-auth==2.2.6
django-admin-thumbnails==0.2.6
django-allauth==0.52.0
django-ckeditor==6.5.1
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-cron==0.6.0
django-crontab==0.7.1
django-currencies==0.10.1
django-embed-video==1.4.8
django-environ==0.8.1
django-extensions==3.2.1
django-heroku==0.3.1
django-humanize==0.1.2
django-imagekit==4.1.0
django-ipware==4.0.2
django-jazzmin==2.6.0
django-modeltranslation==0.17.5
django-money==3.0.0
django-mptt==0.14.0
django-multiselectfield==0.1.12
django-qr-code==3.1.0
django-rest-framework==0.1.0
django-restframework==0.0.1
django-taggit==2.1.0
django-twilio==0.14.3
djangoconvertvdoctopdf==1.0.1
djoser==2.1.0
drf-yasg==1.21.7
env==0.1.0
environ==1.0
environs==8.0.0
exceptiongroup==1.0.0rc9
facebook-sdk==3.1.0
faker==13.13.0
google-api-python-client==2.51.0
graphviz==0.20.1
gunicorn==20.1.0
instaloader==4.9.6
ip2geotools==0.1.6
ipython==8.7.0
ipython-genutils==0.2.0
jaraco.classes==3.2.3
loader==2017.9.11
mysql-connector-python==8.0.31
numerizer==0.2.2
numpy==1.24.1
paycomuz==3.0
paymeuz==0.2
pip-chill==1.0.3
pipenv==2023.2.4
postgres==4.0
public==2020.12.3
pyautogui==0.9.53
pydot==1.4.2
pyopenssl==22.0.0
pysocks==1.7.1
python-decouple==3.6
python-gettext==4.1
python-slugify==6.1.1
python-telegram-bot==13.12
python-twitter==3.5
qrcode==7.3.1
query-string==2020.12.3
reportlab==3.6.12
rest-framework-generic-relations==2.1.0
ruamel.yaml==0.17.21
slugify==0.0.1
stateful-object==0.1
urllib3-secure-extra==0.1.0
uuid==1.30
weasyprint==58.1
wikipedia==1.4.0
xlsxwriter==3.0.9
zopfli==0.2.2
